requirements.txt
Flask==1.1.2
requests==2.25.0
PyJWT==1.7.1
python-jose
